requirements.txt
# Core async and HTTP libraries
aio-pika==9.5.5
aiofiles==24.1.0
aiohttp==3.12.2
aiolimiter==1.1.0
asyncpg==0.30.0
httpx==0.28.1

# Data validation and settings
pydantic==2.11.5
pydantic-settings==2.9.1

# Environment and configuration
python-dotenv==1.1.0

# Blockchain interaction
substrate-interface==1.7.11

# System monitoring
psutil==7.0.0